requirements.txt
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.0
decorator==4.3.2
future==0.17.1
graphviz==0.10.1
kiwisolver==1.0.1
matplotlib==3.0.2
neat-python==0.92
networkx==2.2
numpy==1.16.0
Pillow==5.4.1
pygame==1.9.4
pyglet==1.3.2
pyparsing==2.3.1
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.2
scipy==1.2.0
six==1.12.0
toolz==0.9.0
